first go at moving auth, tests are failing, need to replace solidLogigSingleton
