--- src/authn/authSession.ts
@@ -1,22 +1,0 @@
-import {
-  Session,
-  getClientAuthenticationWithDependencies
-} from '@inrupt/solid-client-authn-browser'
-
-let authSession: Session
-// @ts-ignore
-if (!window.authSession) {
-  authSession = new Session(
-    {
-      clientAuthentication: getClientAuthenticationWithDependencies({})
-    },
-    'mySession'
-  )
-  // @ts-ignore
-  window.authSession = authSession
-} else {
-  // @ts-ignore
-  authSession = window.authSession
-}
-
-export default authSession
