--- test/unit/authn/index.test.ts
@@ -1,11 +1,0 @@
-import { silenceDebugMessages } from '../../helpers/setup'
-import * as Authn from '../../../src/authn/authn'
-import * as Index from '../../../src/authn/index'
-
-silenceDebugMessages()
-
-describe('authn/index', () => {
-  it('exports all of authn/authn', () => {
-    expect(Object.keys(Authn).sort()).toEqual(Object.keys(Index.authn).sort())
-  })
-})
