+++ test/unit/helpers/clearStore.ts
@@ -1,4 +1,4 @@
-import { solidLogicSingleton } from '../../../src/logic'
+import { solidLogicSingleton } from 'solid-logic'
 
 export function clearStore () {
   solidLogicSingleton.clearStore()
